+++ example/lib/page_two.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
 class PageTwo extends StatelessWidget {
-  const PageTwo({Key? key}) : super(key: key);
+  const PageTwo({super.key});
 
   @override
   Widget build(BuildContext context) {